Debugging asynchronous browser calls WIP
Unfortunately Watir does not have the ability to turn redirection off.  Because of this I'm having some trouble implementing the scraper to search for a band correctly.  When you type in a band that only has one entry in the metal-archives database, the search timeline goes through three seperate hits.

1: You get the search band page with the html table that initially says it's searching.
2: There's a moment where the ajax returns something but the redirection hasn't happened yet.
3: Redirection to the band's profile page.

Because I cannot figure out a way to see if any divs have been removed or changed in the moment for step two, I'm not sure how to work around this besides making watir just timeout for a few seconds while metal-archives does its magic.  That's obviously not a very efficient route to take, it feels really hacky.