More progress wip
Stil having the same issue with dealing with the ajax response.  This code is closer but now I'm running into an issue where sometimes Watir will throw an exception about the table not being found.  This is strange since it only appears sometimes.  Not sure how to work around this.