Add associated bands for other case, found edge case bug
So now there's an issue with adding associated bands if a member of the band has no associated bands.  Then the array of band members attributes start moving around and getting passed incorrectly.  The only way to fix this problem I'm seeing is to start refactoring down this giant chunked if else statement and start extracting that logic into it's own methods.  Once I do that it will become more clear what needs to be the if else statment that checks if the associated bands of a member exisits.